+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Add request body parsing for POST data
 * Add handlers module for convenient error response handling
 * Add isoToLocale util function
+* Expose all static resources instead of individual images, scripts, and styles
 
 ### 0.0.4
 * Upgrade all deps to latest, Express to v2.5.9